requirements.txt
atpublic==3.1.1
duckdb==0.7.1
duckdb-engine==0.6.9
ibis-framework==4.1.0
markdown-it-py==2.2.0
mdurl==0.1.2
multipledispatch==0.6.0
numpy==1.24.2
pandas==1.5.3
parsy==2.1
py4j==0.10.9.5
pyarrow==10.0.1
Pygments==2.14.0
pyspark==3.3.2
python-dateutil==2.8.2
pytz==2022.7.1
rich==13.3.2
six==1.16.0
SQLAlchemy==1.4.46
sqlalchemy-views==0.3.2
sqlglot==10.6.4
toolz==0.12.0
typing_extensions==4.5.0
